src/api_client.py
"""
HTTP client for barcodes.dev API.

This module handles all communication with the barcode generation API,
including request formatting, response parsing, and error handling.
"""

import base64
import io
import logging
from typing import Dict, List, Optional, Any
from dataclasses import dataclass

import requests
from PIL import Image

logger = logging.getLogger(__name__)


@dataclass
class BarcodeResult:
    """Result from barcode generation API."""
    image: Image.Image
    metadata: Dict[str, Any]
    format: str = "png"


class BarcodeAPIClient:
    """Client for barcodes.dev barcode generation API."""

    def __init__(
        self,
        base_url: str = "http://localhost:5001",
        timeout: int = 30,
        retry_attempts: int = 3
    ):
        """
        Initialize API client.

        Args:
            base_url: Base URL of the barcodes.dev API
            timeout: Request timeout in seconds
            retry_attempts: Number of retry attempts on failure
        """
        self.base_url = base_url.rstrip('/')
        self.timeout = timeout
        self.retry_attempts = retry_attempts
        self.session = requests.Session()

    def generate_barcode(
        self,
        barcode_type: str,
        text: str,
        degradation: Optional[Dict] = None
    ) -> Optional[BarcodeResult]:
        """
        Generate a barcode with optional degradation.

        Args:
            barcode_type: Symbology type (code128, qr, upca, etc.)
            text: Data to encode in barcode
            degradation: Optional degradation configuration

        Returns:
            BarcodeResult with image and metadata, or None on failure
        """
        # TODO: Implement API call
        raise NotImplementedError("API client not yet implemented")

    def get_degradation_presets(self) -> List[Dict]:
        """
        Get available degradation presets from API.

        Returns:
            List of preset configurations
        """
        # TODO: Implement API call
        raise NotImplementedError("API client not yet implemented")

    def get_supported_symbologies(self) -> List[str]:
        """
        Get list of supported barcode symbologies.

        Returns:
            List of symbology names
        """
        # TODO: Implement API call
        raise NotImplementedError("API client not yet implemented")

    def health_check(self) -> bool:
        """
        Check if API server is reachable.

        Returns:
            True if server is healthy, False otherwise
        """
        try:
            response = self.session.get(
                f"{self.base_url}/health",
                timeout=5
            )
            return response.status_code == 200
        except requests.RequestException:
            return False

    def _decode_image(self, base64_data: str) -> Image.Image:
        """Decode base64 image data to PIL Image."""
        image_data = base64.b64decode(base64_data)
        return Image.open(io.BytesIO(image_data))
