requirements.txt
# HTTP client
requests>=2.28.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0

# Configuration
PyYAML>=6.0

# CLI
click>=8.0.0

# Progress bars
tqdm>=4.64.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
responses>=0.22.0  # Mock HTTP responses

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
